% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{acfpacf_plot}
\alias{acfpacf_plot}
\title{(P)ACF plots}
\usage{
acfpacf_plot(data, lagmax = length(data), mfrow = c(2, 1), ...)
}
\arguments{
\item{data}{numeric ector containing}

\item{laxgmax}{numeric value containing the maximum number of lags to display}
}
\value{
Page of both ACF and PACF until lagmax
}
\description{
Produce plots of both ACF and PACF for a
given maximum number of lags.
}
\details{
Reduce length of code as this operation should be made several times.
This will be implemented in \code{ggplot} if shown in the finalo thesis.
}
\examples{
acfpacf_plot(max_years$data)
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
