% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapFunc.R
\name{ciBootFunc}
\alias{ciBootFunc}
\alias{mle.ksiFun}
\alias{sdmle.ksiFun}
\alias{ciBootFunc}
\alias{ci2Boot.tFunc}
\title{MLE's GPD estimator for \eqn{\xi}
and boostrap evaluations}
\usage{
mle.ksiFun(x)

sdmle.ksiFun(x)

ciBootFunc(x, B, level = 0.05, print = T)

ci2Boot.tFunc(x, B1, B2, level = 0.05, print = T)
}
\arguments{
\item{x}{A numeric vector of data to be fitted.}

\item{B}{number of bootstrap replicates for simple bootstrap ci}

\item{level}{confidence level for the confidence intervals}

\item{B1}{number of bootstrap replicates for double bootstrap ci (2)
for the first stage}

\item{B1}{number of bootstrap replicates for double bootstrap ci (2)
for the second stage}
}
\value{
MLE computed for the GPD or the standard deviation of this MLE
}
\description{
Function to compute directly our Maximum Likelihood Estimator (MLE) for
the shape parameter of the Generalized Pareto Distribution.
It isalso possible to determine bootstrapped confidence interals,
either by simple or by double bootstrap
}
\details{
This function is useful to decrease the amount of code and compute directly
these estimates.
}
\examples{
X <- above_ured$exceed_red
ksi.chap <- gpd_varu_red$mle[2]
sd.ksi.chap <- sdmle.ksiFun(X)
c500 <- ciBootFunc(B = 500)   # 9 sec.
cc <- ci2Boot.tFunc(B1=100,B2=34)
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
