% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{diplot_fast}
\alias{diplot_fast}
\title{Dispersion Index Plot for threshld selection}
\usage{
diplot_fast(data, u.range, nt = max(200, nrow(data)))
}
\description{
Rearranging diplot() function from \code{POT} package
because there were speed problems. SEE DETAILS on the \code{POT}'s package
}
\examples{
data.di <- data.frame(time = seq(1:length(max_all)), obs = max_all)
events <- clust(data.di, u = 26, tim.cond = 20/365, clust.max = TRUE)
diplot(events[,-3], u.range = c(28,36) ) # too slow
diplot_fast(events[,-3], u.range = c(25,35), nt = 1000 )  # Better
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
