% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{func_season}
\alias{func_season}
\title{Retrieve seasons from months}
\usage{
func_season(month)
}
\arguments{
\item{month}{of the year from which we want to return the season}
}
\value{
the meteoroglogical season associated to each month
}
\description{
This function aims to retrieve specific seasons from a month standing
and the temperatures in terms of seasons
}
\examples{
# Retrieve seasons based on meteorological seasons

TXTN_closed$season <- sapply(TXTN_closed$month, function(x) func_season(x))

}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
