% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{genHill}
\alias{genHill}
\title{Generalized Hill estimates}
\usage{
genHill(data, gamma, plot = FALSE, add = FALSE, ...)
}
\arguments{
\item{data}{numeric vector of extremes}
}
\value{
list of K and \eqn{\gamma} (Zipf estimates)
}
\description{
Computes the Hill estimates of gamma (Section 5.2.3)
for a numeric vector of observations (data) and as a function of k
If plot=TRUE then the estimates are plotted as a function of k
If add=TRUE then the estimates are added to an existing plot
}
\details{
Based on the functions provided in \url{http://lstat.kuleuven.be/Wiley/}.
Not used as it holds only for fat-tailed distributions (\eqn{\xi>0})
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
