% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{gev.fit2}
\alias{gev.fit2}
\title{gev.fit() from ismev redefined}
\usage{
gev.fit2(xdat, ydat = NULL, mul = NULL, sigl = NULL, shl = NULL,
  mulink = identity, siglink = identity, shlink = identity,
  muinit = NULL, siginit = NULL, shinit = NULL, show = TRUE,
  method = "Nelder-Mead", maxit = 10000, browser = F, solve.tol = 1e-18,
  ...)
}
\arguments{
\item{solve.tol}{and those found in ismev : ?ismev::gev.Fit}
}
\description{
Redifined function from ismev because there were errors when
computing the cubic model for the location parameter in a nonstationary model.
We needed to insert a parameter allowing to control the solerance when needed
to solve(x$hessian).
}
\examples{
## 'Cubic' model ?
ti3 <- matrix(ncol = 3, nrow = length(max_years$data))
ti3[ ,1] <- seq(1, length(max_years$data), 1)
ti3[ ,2] <- (ti3[,1])^2
ti3[ ,3] <- (ti3[,1])^3
gev_nonstatio3 <- gev.fit(max_years$data/1000, ydat = ti3, mul = c(1, 2, 3))
gev_nonstatio3 <- PissoortThesis::gev.fit2(max_years$data, ydat = ti3,
                                          mul = c(1, 2, 3),
                                          browser = T, solve.tol = 1e-25)
# System is singular if we do not scale the data.  But if we scale, the are still
#NaNs produced in the covariance matrix.
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
