% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{gev.nloglik}
\alias{gev.nloglik}
\title{Return Levels with nonstationarity}
\usage{
gev.nloglik(mu, sig, xi, data)
}
\arguments{
\item{data}{numeric vector containing the GEV block-maxima}

\item{gev_nstatio}{Nonstationary GEV fitted model of class \code{gev.fit}
(from package \code{ismev})}

\item{t}{Maximum time period for which the return levels are considered}

\item{m}{Return period}
}
\value{
The return levels for the considered time period (t)
}
\description{
Compute return levels plot of nonstationary model with the data (in years)
}
\examples{
rl_10_lin <- return.lvl.nstatio(max_years$df$Year,
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
