% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{gev.proflik}
\alias{gev.proflik}
\title{GEV profile neg profile likelihood}
\usage{
gev.proflik(a)
}
\arguments{
\item{a}{is a two dimensionnal vector containing the scale
and the shape parameter respectively}
}
\value{
the negative profile ikelihood
}
\description{
# computes the negative profile log-likelihood
(credits to E. Gilleland and ismev package)
}
\examples{
for(i in 1:nint) {
xp <- x[i]
opt <- optim(sol, gev.proflik)
sol <- opt$par ; v[i] <- opt$value
}
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
