% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{chains.plotOwn}
\alias{chains.plotOwn}
\alias{mixchains.Own}
\title{Plots to assess the mixing of the Chains}
\usage{
chains.plotOwn(data, ..., title = "TracePlots of the generated Chains ")

mixchains.Own(data, moreplot = F,
  title = "TracePlots of the generated Chains ")
}
\arguments{
\item{data}{numeric vector containing the GEV block-maxima}

\item{...}{Other parameters from \code{gridExtra::grid.arrange()}}

\item{title}{Global title for the plot}
}
\value{
a grid.arrange() of ggplots.
}
\description{
Compute the ggplots for each parameter of interest in a single page.
}
\examples{
data("max_years")
fn <- function(par, data) -log_post0(par[1], par[2], par[3], data)
param <- c(mean(max_years$df$Max),log(sd(max_years$df$Max)), 0.1 )
# opt <- optim(param, fn, data = max_years$data,
#              method="BFGS", hessian = TRUE)
opt <- nlm(fn, param, data = max_years$data,
           hessian=T, iterlim = 1e5)
start <- opt$estimate
Sig <- solve(opt$hessian)
ev <- eigen( (2.4/sqrt(2))^2 * Sig)
varmat <- ev$vectors \%*\% diag(sqrt(ev$values)) \%*\% t(ev$vectors)
# (MH)
set.seed(100)
mh.mcmc1 <- MH_mcmc.own(start, varmat \%*\% c(.1,.3,.4))
mh.mcmc1$mean.acc_rates

chains.plotOwn(mh.mcmc1$out.chain)

# (GIBBS)
# k chains with k different starting values
set.seed(100)
gibbs.trend <- gibbs.trend.own(start, propsd = c(.5, 1.9, .15, .12),
                              iter = 1000)
## TracePlots
chain.mix <- cbind.data.frame(gibbs.trend$out.chain,
                             iter.chain = rep(1:500, 4))
mixchains.Own(chain.mix)
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
