% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{gibbssig3}
\alias{gibbssig3}
\alias{log_post3}
\alias{gibbs.trend.sig3own}
\title{Return Levels with nonstationarity}
\usage{
log_post3(mu0, mu1, logsig0, logsig1, xi, data, model.mu = mu0 + mu1 * tt,
  mnpr = c(30, 0, 1, 0, 0), sdpr = c(10, 40, 10, 10, 10), ic = F)

gibbs.trend.sig3own(start, propsd = c(0.5, 2.5, 2, 0.08, 0.08), iter = 1000,
  data = max_years$data)
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
