% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{grid_arrange_legend}
\alias{grid_arrange_legend}
\title{Format legend for several ggplots}
\usage{
grid_arrange_legend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"))
}
\value{
A \code{grid.arrange()} with legend formatted
}
\description{

}
\details{
This function is useful to decrease the amount of code for each ggplots
generated as this thesis will use exclusive \code{ggplot2} for the plots.
Values of other parameters such as colours could be changed inside the function.
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
