% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{gev.nloglik}
\alias{gev.nloglik}
\alias{log_post0}
\title{Negative GEV log-likelihood and Log-Posterior with
uninformative normal priors.}
\usage{
gev.nloglik(mu, sig, xi, data)

log_post0(mu, logsig, xi, data)
}
\arguments{
\item{mu}{numeric representig the location parameter of the GEV}

\item{sig}{or \code{logsig} are numeric representig the scale parameter
of the GEV. BE CAREFULL : this is in logarithm for the log_post0 only in
order to avoid computational problems later in the MCMC's.}

\item{xi}{numeric representig the shape parameter of the GEV}

\item{data}{numeric vector representing the data (GEV) of interest.}
}
\value{
a numeric value representing the negative log-likelihood or the log-posterior
of interest.
}
\description{
Functions to compute the negative log-likelihood of a GEV
distribution. Whereas this can be used for other purpose, this
also allows in particular to compute the log-posterior with diffuse normal
priors. Note that we could add parameters to control the informativness of
the priors, but as we have no reliable information, we decide to arbitrarily
 fix it to large values, to improve computation . (More parameters are
 harmful for computation time)
}
\examples{
data('max_years')
# Optimize the log-Posterior Density Function to find starting values
fn <- function(par, data) -log_post0(par[1], par[2], par[3], data)
param <- c(mean(max_years$df$Max),log(sd(max_years$df$Max)), 0.1 )
opt <- nlm(fn, param, data = max_years$data,
           hessian=T, iterlim = 1e5)
start <- opt$estimate

}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
