% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{log_post1}
\alias{log_post1}
\title{Return Levels with nonstationarity}
\usage{
log_post1(mu0, mu1, logsig, xi, data, model.mu = mu0 + mu1 * tt,
  mnpr = c(30, 0, 0, 0), sdpr = c(40, 40, 10, 10))
}
\arguments{
\item{data}{numeric vector containing the GEV block-maxima}

\item{gev_nstatio}{Nonstationary GEV fitted model of class \code{gev.fit}
(from package \code{ismev})}

\item{t}{Maximum time period for which the return levels are considered}

\item{m}{Return period}
}
\value{
The return levels for the considered time period (t)
}
\description{
Compute return levels plot of nonstationary model with the data (in years)
}
\examples{
rl_10_lin <- return.lvl.nstatio(max_years$df$Year,

}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
