% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{MeanExcess}
\alias{MeanExcess}
\title{Mean Excess function and plots}
\usage{
MeanExcess(data, plot = FALSE, k = FALSE)
}
\arguments{
\item{data}{numeric vector of extremes}

\item{plot}{specify if we want a plot or not}

\item{k}{specifies if we want the plots as a function of k (=TRUE)
or as a function of the order statistics \eqn{X_{n-k,n}} (=FALSE)}
}
\value{
List containing the values of k, the corresponding order statistics
\eqn{X_{n-k}} and the mean excess scores associated \code{e_{k,n}}
}
\description{
Computes the mean excess scores \eqn{e_{k,n}} for a numeric vector of data
}
\details{
Based on the functions provided in \url{http://lstat.kuleuven.be/Wiley/}.
If plot=T, the mean excess scores are plotted as a function of k
If plot=TRUE and k=TRUE then the mean excess scores are
plotted as a function of the order statistics X_n-k,n
}
\examples{
max_all <- TXTN_closed$TX
MeanExcess(max_all, plot = T, k =1000)
}
\references{
Jan Beirlant, Yuri Goegebeur, Johan Segers, and Jozef Teugels.
Statistics of Extremes: Theory and Applications. John Wiley & Sons, March 2006.
Section \eqn{\boldsymbol{1.2.2}}
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
