% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{PissoortThesis-package}
\alias{PissoortThesis}
\alias{PissoortThesis-package}
\title{Store the useful functions used during the thesis of Antoine Pissoort.
Package building has been made with the use of \code{roxygen2}
the functions are divided in several blocks
\itemize{
\item 'Standard' (frequentist) Extreme Value Analysis, both stationary
and nonstationary.
\item Neural Netowrks
\item bootstrap evaluation and performance comparisons (check if we include it)
\item Bayesian analysis
}}
\description{
Store the useful functions used during the thesis of Antoine Pissoort.
Package building has been made with the use of \code{roxygen2}
the functions are divided in several blocks
\itemize{
\item 'Standard' (frequentist) Extreme Value Analysis, both stationary
and nonstationary.
\item Neural Netowrks
\item bootstrap evaluation and performance comparisons (check if we include it)
\item Bayesian analysis
}
}
\references{
Jan Beirlant, Yuri Goegebeur, Johan Segers,
 and Jozef Teugels. Statistics of Extremes: Theory and Applications.
 John Wiley & Sons, March 2006.

Stuart Coles. An Introduction to Statistical Modeling of Extreme Values.
 Springer Series in Statistics. Springer London, London, 2001.

Alex J. Cannon. A flexible nonlinear modelling framework for
nonstationary generalized extreme value analysis in hydroclimatology.
 Hydrological Processes, 24(6):673–685, March 2010.

Dipak K. Dey and Jun Yan. Extreme Value Modeling and Risk Analysis:
 Methods and Applications. CRC Press, January 2016.

Marcelo Hartmann and Ricardo Ehlers. Bayesian Inference for
Generalized Extreme Value Distributions via Hamiltonian Monte Carlo.
Communications in Statistics - Simulation and Computation, pages 0–0, March 2016.
}
