% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{posterior_pred_ggplot}
\alias{posterior_pred_ggplot}
\title{Predictive Posterior density plots}
\usage{
posterior_pred_ggplot(Data = max_years$df,
  Model_out.chain = gibbs.trend$out.chain, from = 1,
  until = nrow(max_years$df), n_future = 0, by = 10, x_coord = c(27,
  35))
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
