% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{Pred_Dens_ggPlot}
\alias{Pred_Dens_ggPlot}
\title{Predictive Posterior density plots}
\usage{
Pred_Dens_ggPlot(year, data_ppd, ylim = c(0.01, 0.55), xlim = c(27.5, 36))
}
\arguments{
\item{year}{numeric value giving the year for which we want to compute the
predictive density}

\item{data_ppd}{numeric matrix or df of size [n,p] containing the posterior predictive
samples, where \code{n} is the number of simulated samples and \code{p}
 is the prediction horizon.}

\item{ylim}{and \code{xlim} define the grid for the plot.}
}
\value{
a ggplot
}
\description{
Compute a \code{ggplot} of the Density associated with the
Posterior Predictive Distribution (PPD)
}
\examples{
# For the PPD density plot of the year 2026 :
gg4 <- PissoortThesis::Pred_Dens_ggPlot(2026, repl2)
# where repl2 is the matrix containing the predictive samples (see Bayes_own_gev.R)

}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
