% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{pred_post_samples}
\alias{pred_post_samples}
\title{Predictive Posterior samples}
\usage{
pred_post_samples(data = max_years$df,
  model_out.chain = gibbs.trend$out.chain, from = 1,
  until = nrow(max_years$df), n_future = 0, seed = NULL)
}
\description{
Compute posterior predictive samples from the obtained model (gibbs.trend)
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
