% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{return.lvl.nstatio}
\alias{return.lvl.nstatio}
\title{Return Levels with nonstationarity}
\usage{
return.lvl.nstatio(data, gev_nstatio, start = max(data), t = 100, m = 10,
  plot = T, out = F, ...)
}
\arguments{
\item{data}{numeric vector containing the GEV block-maxima}

\item{gev_nstatio}{Nonstationary GEV fitted model of class \code{gev.fit}
(from package \code{ismev})}

\item{t}{Maximum time period for which the return levels are considered}

\item{m}{Return period}

\item{plot}{returns a ggplot of the return levels}

\item{out}{outputs the return levels (TRUE)}

\item{...}{other parmaters to be passed to geom_point() of \code{ggplot}}
}
\value{
a ggplot in the \code{g} object
 and  The return levels for the considered time period (t) in \code{rl}.
}
\description{
Compute return levels plot of nonstationary model with the data (in years)
}
\examples{
rl_10_lin <- return.lvl.nstatio(max_years$df$Year,
                                gev_nonstatio, t = 500, m = 10)
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
