% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{rl_piss}
\alias{rl_piss}
\alias{gev.rl.gradient}
\alias{rl_piss}
\title{Return Levels with ggplot2}
\usage{
gev.rl.gradient(est, p)

rl_piss(est, mat, data, thm = theme_piss())
}
\arguments{
\item{est}{numeric vector containing the GEV estimates of the 3 parameters.}

\item{p}{the given probability of exceeding the threshold in a year. Only used
through \code{gev.rl.gradient} function.}

\item{mat}{Covariance matrix of the estimates}

\item{data}{numeric vector containing the GEV block-maxima}

\item{theme}{ggplot2's theme used for the plots}
}
\value{
Corresponding plot of the return levels
}
\description{
\code{rl_piss()} is designed to compute plots from \code{ggplot2}
package and \code{gev.rl.gradient()} is used inside \code{rl_piss()} to compute
the gradient estimates.
}
\details{
See \code{ismev} package details for more in-depth informations
}
\examples{
rl_piss(gev_tx$mle, gev_tx$cov, gev_tx$data)
}
\references{
Stuart Coles. An Introduction to Statistical Modeling of
 Extreme Values. Springer Series in Statistics. Springer London, London, 2001.
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
