% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{return_levels_gg}
\alias{return_levels_gg}
\alias{rl.post_gg}
\alias{rl.pred_gg}
\title{Return Levels with nonstationarity}
\usage{
rl.post_gg(post, npy, method = c("gev", "gpd"), ci = 0.9, ...)

rl.pred_gg(post, qlim, npy, method = c("gev", "gpd"), period = 1, ...)
}
\arguments{
\item{data}{numeric vector containing the GEV block-maxima}

\item{gev_nstatio}{Nonstationary GEV fitted model of class \code{gev.fit}
(from package \code{ismev})}

\item{t}{Maximum time period for which the return levels are considered}

\item{m}{Return period}
}
\value{
The return levels for the considered time period (t)
}
\description{
Compute return levels plot of nonstationary model with the data (in years)
}
\examples{
rl_10_lin <- return.lvl.nstatio(max_years$df$Year,

}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
