% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Shiny application}
\usage{
runExample(example)
}
\arguments{
\item{example}{chooses the example you want. Run runExample() to have the list}
}
\examples{
library(PissoortThesis)
runExample('Name of the app') # and you will see the app you want
# You juste have to run runExample('') with whatever incorrect name to display the names
# of the existing apps

}
