% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{theme_piss}
\alias{theme_piss}
\title{Homogeneous theme for ggplots}
\usage{
theme_piss(size_p = 18, size_c = 14, size_l = 12, theme = theme_bw(),
  ...)
}
\arguments{
\item{size_p}{Size of the plot's title.}

\item{size_c}{Size of the axis' title.}

\item{size_l}{Size of the legend's title.}

\item{theme}{ggplot's theme for the plot. Set it to NULL if you 
want put your own theme construction in the "..."}
}
\value{
A personalized ggplot2 theme object to add to every builded plots.
}
\description{
Theme function for our builded \code{ggplots}.
Useful get coherent and similar colours, themes,... for all plots.
Also useful to decrease the number of
}
\details{
This function is useful to decrease the amount of code for each ggplots
generated as this thesis will use exclusive \code{ggplot2} for the plots.
Values of other parameters such as colours could be changed inside the function.
}
\examples{
# To generate the PP-plot ! See code for more details
ggplot(data = data.frame(empirical,model_est), aes(x=empirical,y=model_est)) +
geom_point(shape = 1, col = "#33666C") + geom_abline(intercept=0,slope=1,col="red") +
theme_piss() +  ggtitle("Probability plot")

}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
