% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{yearly.extrm}
\alias{yearly.extrm}
\title{Yearly extrema retrieval}
\usage{
yearly.extrm(list.y = list_by_years, Fun = max, tmp = "TX")
}
\arguments{
\item{list.y}{List containing numeric vectors of (daily typically) extrema
for each year. Length of \code{list.y} is thus the number of years.}

\item{Fun}{which function to compute, thus here typically max() or min()}

\item{tmp}{Specify if we want 'TX' (max. temp.) or 'TN' (min. temp.)}
}
\value{
List of 2 elements : \code{df} which is a data.frame used
to compute ggplots. \code{data} is the vector of yearly extrema.
}
\description{
retrieve the (yearly) extremes for a given list of extrema.
The two last arguments specify if we want rather minima (Fun=min, tmp='TN')
or maxima. This returns the data AND data.frame with the corresponding year.
}
\details{
Easier to handle for ggplots
}
\examples{
# For maxima (default)
max_years <- yearly.extrm()
# For minnima
min_years <- yearly.extrm(Fun = min, tmp = "TN")
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
