% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFunc.R
\name{gibbstrend3}
\alias{gibbstrend3}
\alias{log_post3}
\alias{gibbs.trend.sig3own}
\title{Return Levels with nonstationarity}
\usage{
log_post3(mu0, mu1, sig0, sig1, xi, data, model.mu = mu0 + mu1 * tt,
  mnpr = c(30, 0, 1, 0, 0), sdpr = c(10, 40, 10, 10, 10))

gibbs.trend.sig3own(start, propsd = c(0.5, 2.5, 2, 0.08, 0.08), iter = 1000,
  data = max_years$data)
}
\arguments{
\item{data}{numeric vector containing the GEV block-maxima}

\item{gev_nstatio}{Nonstationary GEV fitted model of class \code{gev.fit}
(from package \code{ismev})}

\item{t}{Maximum time period for which the return levels are considered}

\item{m}{Return period}
}
\value{
The return levels for the considered time period (t)
}
\description{
Compute return levels plot of nonstationary model with the data (in years)
}
\examples{
rl_10_lin <- return.lvl.nstatio(max_years$df$Year,

}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
