% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R
\name{return.lvl.nstatio}
\alias{return.lvl.nstatio}
\title{Return Levels with nonstationarity}
\usage{
return.lvl.nstatio(data, gev_nstatio, t = 100, m = 10)
}
\arguments{
\item{data}{numeric vector containing the GEV block-maxima}

\item{gev_nstatio}{Nonstationary GEV fitted model of class \code{gev.fit}
(from package \code{ismev})}

\item{t}{Maximum time period for which the return levels are considered}

\item{m}{Return period}
}
\value{
The return levels for the considered time period (t)
}
\description{
Compute return levels plot of nonstationary model with the data (in years)
}
\examples{
rl_10_lin <- return.lvl.nstatio(max_years$df$Year,
                                gev_nonstatio, t = 500, m = 10)
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
