% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1UsedFunc.R, R/BayesFunc.R
\name{rl_piss}
\alias{rl_piss}
\alias{gev.rl.gradient}
\alias{rl_piss}
\alias{return_levels_gg}
\alias{rl.post_gg}
\alias{rl.pred_gg}
\title{Return Levels with ggplot2}
\usage{
gev.rl.gradient(est, p)

rl_piss(est, mat, data, thm = theme_piss())

rl.post_gg(post, npy, method = c("gev", "gpd"), ci = 0.9, ...)

rl.pred_gg(post, qlim, npy, method = c("gev", "gpd"), period = 1, ...)
}
\arguments{
\item{est}{numeric vector containing the GEV estimates of the 3 parameters.}

\item{p}{the given probability of exceeding the threshold in a year. Only used
through \code{gev.rl.gradient} function.}

\item{mat}{Covariance matrix of the estimates}

\item{data}{numeric vector containing the GEV block-maxima}

\item{theme}{ggplot2's theme used for the plots}

\item{data}{numeric vector containing the GEV block-maxima}

\item{gev_nstatio}{Nonstationary GEV fitted model of class \code{gev.fit}
(from package \code{ismev})}

\item{t}{Maximum time period for which the return levels are considered}

\item{m}{Return period}
}
\value{
Corresponding plot of the return levels

The return levels for the considered time period (t)
}
\description{
\code{rl_piss()} is designed to compute plots from \code{ggplot2}
package and \code{gev.rl.gradient()} is used inside \code{rl_piss()} to compute
the gradient estimates.

Compute return levels plot of nonstationary model with the data (in years)
}
\details{
See \code{ismev} package details for more in-depth informations
}
\examples{
rl_piss(gev_tx$mle, gev_tx$cov, gev_tx$data)
rl_10_lin <- return.lvl.nstatio(max_years$df$Year,

}
\references{
Stuart Coles. An Introduction to Statistical Modeling of
 Extreme Values. Springer Series in Statistics. Springer London, London, 2001.
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}

Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
